+++ src/handlers/projectActionHandler.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import * as path from "path";
-import { PromptTemplate } from "../notion/parser";
+import { PromptTemplate } from "../knowledge/parser";
 import { log, logError } from "../logger";
 import { resolveModel } from "../utils/modelResolver";